// src/api/interface/config/IInput.ts
/*
* This program and the accompanying materials are made available under the terms of the
* Eclipse Public License v2.0 which accompanies this distribution, and is available at
* https://www.eclipse.org/legal/epl-v20.html
*
* SPDX-License-Identifier: EPL-2.0
*
* Copyright Contributors to the Zowe Project.
*
*/

export type InputType = "prompt" | "env" | "user";
export interface IInput {
    sources: InputType[];
    allowBlank?: boolean;
    mask?: boolean;
    desc?: string;
}